rename: util 함수 makeRandomNumber를 makeRandomInRange로 파일명 변경
